New structure for CV created
