+++ Javascripts/cv/cv_html_loader.js
@@ -1,7 +1,7 @@
 const main = document.querySelector('main');
 const styleDiv = document.getElementById('stylesheetContainer');
 
-export function cvHtmlLoader() {
+export function cvHtmlLoad() {
     styleDiv.innerHTML = `<link rel="stylesheet" href="./Stylesheets/cv.css">`;
     main.innerHTML = `<div class="aboutMe-div flexbox">
     <p class="aboutMe-titel agency">Jan Lehner</p>
