+++ Javascripts/cv/loadCV.js
@@ -1,35 +1,9 @@
-import { cvHtmlLoader } from "./cv_html_loader.js";
+import { cvHtmlLoad } from "./cv_html_loader.js";
 import { cvClicked } from "../navBtnSelector.js";
+import { cvObserve } from "./cv_observer.js";
 
 export function loadCV() {
-    cvHtmlLoader();
+    cvHtmlLoad();
     cvClicked();
-    const observerLines = new IntersectionObserver((entries) => {
-        entries.forEach((entry) => {
-            if (entry.isIntersecting) {
-                entry.target.classList.add('showLine');
-            } else {
-                entry.target.classList.remove('showLine');
-            }
-        });
-    });
-
-    const observerFade = new IntersectionObserver((entries) => {
-        entries.forEach((entry) => {
-            if (entry.isIntersecting) {
-                entry.target.classList.add('showFade');
-            } else {
-                entry.target.classList.remove('showFade');
-            }
-        });
-    });
-
-    const hiddenLines = document.querySelectorAll('.hiddenLine');
-    hiddenLines.forEach((el) => observerLines.observe(el));
-
-    const hiddenFadeFR = document.querySelectorAll('.hiddenFadeFR');
-    hiddenFadeFR.forEach((el) => observerFade.observe(el));
-
-    const hiddenFadeFL = document.querySelectorAll('.hiddenFadeFL');
-    hiddenFadeFL.forEach((el) => observerFade.observe(el));
+    cvObserve();
 };
